+++ bonus - testovani/kamen_nuzky_papir.py
@@ -1,3 +1,5 @@
+# file (or module) which contains the whole game
+
 from random import choice
 
 slova = ["kámen", "nůžky", "papír"]
